move static function to new utils file
